console: Extract device utils to lib
